test(canon): split release evidence and divergence report modules
